% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bottomcode}
\alias{bottomcode}
\title{Bottom-code variables}
\usage{
bottomcode(x, value)
}
\arguments{
\item{x}{variable}

\item{value}{Lower bound}
}
\description{
Censor a variable below a certain lower bound
}
