% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{bwplot}
\alias{bwplot}
\title{Scatterplot with black-and-white style}
\usage{
bwplot(
  x,
  y = NULL,
  ...,
  type = "p",
  hlines = NULL,
  hlines.col = "lightgray",
  hlines.lty = 1,
  hlines.lwd = 1,
  vlines = NULL,
  vlines.col = "lightgray",
  vlines.lty = 1,
  vlines.lwd = 1,
  xat = NULL,
  yat = NULL,
  axis.col = "lightgray",
  bgcolor = "gray99",
  pch = 21,
  bg = "lightblue",
  col = "black",
  v_over_h = FALSE
)
}
\arguments{
\item{x}{Coordinates of points in the plot}

\item{y}{Coordinates of points in the plot (optional)}

\item{...}{Optional graphics arguments}
}
\value{
None.
}
\description{
Scatterplot with black-and-white style
}
