% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefplot.R
\name{coefplot}
\alias{coefplot}
\title{Coefficient plot}
\usage{
coefplot(object, omit = NULL, horizontal = TRUE, labels = TRUE, ...)
}
\arguments{
\item{omit}{string with regular expression to omit coefficients}

\item{horizontal}{estimates are plotted on the x-axis (default) or the y-axis}

\item{x}{object with fitted model}
}
\description{
Plots estimates and coefficients from fitted object using \code{coef(summary())}
}
