% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ilogit}
\alias{ilogit}
\title{Inverse logistic function = exp(x) / (1 + exp(x))}
\usage{
ilogit(x)
}
\arguments{
\item{x}{real number}
}
\description{
Inverse logistic function = exp(x) / (1 + exp(x))
}
\examples{
curve(ilogit, from=-5, to=5)
}
