% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{Impute missing values}
\usage{
impute(x, value)
}
\arguments{
\item{x}{vector}

\item{value}{replacement value (="median" will impute the median of the non-missing values)}
}
\description{
Replace missing values with given value
}
\examples{
impute(c(NA, 12, NA, 1:10), 0)
}
