% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{percent}
\alias{percent}
\title{Print proportions as percentages}
\usage{
percent(x, digits = 0, brackets = FALSE)
}
\arguments{
\item{x}{real number (proportion)}
}
\description{
Transforms proportions in strings with percent sign (with or without brackets).
}
\examples{
p <- 0.15
percent(0.15)
}
