% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dens.R
\name{plot_dens}
\alias{plot_dens}
\title{Plot kernel density curve with black and white style with shaded area}
\usage{
plot_dens(x, rescale = TRUE, col.area = "dodgerblue", ylab = "density", ...)
}
\arguments{
\item{x}{it can be the data from which the estimate is to be computed, or an 'density' object}

\item{rescale}{if TRUE, ensures density stays in \link{0,1}; default true.}

\item{col.area}{color of the shaded area; default "dodgerblue".}

\item{...}{Optional graphics arguments}
}
\value{
None.
}
\description{
Plot kernel density curve with black and white style with shaded area
}
