% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_latex.R
\name{preview_latex}
\alias{preview_latex}
\title{Preview table in PDF using pandoc}
\usage{
preview_latex(x, filename = NULL, pandoc.opt = NULL, header = NULL)
}
\arguments{
\item{x}{vector of well-formatted latex table (e.g., from kable)}

\item{filename}{name of destination file (optional)}

\item{pandoc.opt}{character with pandoc options (-V <variables> and -H <header>)}
}
\description{
Converts a table in latex into a PDF document (class standalone)
}
\examples{
tab <- xtabs( ~ cyl + vs, data = mtcars) 
preview_latex(knitr::kable(tab, format = "latex"))
}
