% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Save plot to file}
\usage{
save_plot(FUN, file = NULL, w = 6, h = 6, view = TRUE, ...)
}
\arguments{
\item{FUN}{Function making the plots (e.g., hist)}

\item{file}{Where to save the figure}

\item{w}{Figure width in inches}

\item{h}{Figure height in inches}

\item{view}{Logical, open the saved file}

\item{...}{Arguments passed to the FUN function}
}
\description{
After having defined a plot function, use this tool to save the plot to file (.png or .pdf).
}
\examples{
save_plot(hist, x = rnorm(1e3), file = tempfile(fileext = ".png"))
}
