% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_replace}
\alias{str_replace}
\title{Pipe-friendly version of gsub function}
\usage{
str_replace(x, pattern, replace, ...)
}
\arguments{
\item{x}{strings}

\item{pattern}{regex strings}

\item{replace}{replacement}

\item{...}{other parameters to gsub}
}
\description{
Performs replacement of patterns.
}
