% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{topcode}
\alias{topcode}
\title{Top-code variables}
\usage{
topcode(x, value)
}
\arguments{
\item{x}{Variable to be censored}

\item{value}{Upper bound}
}
\description{
Censor a variable above a certain upper bound
}
