% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{wp}
\alias{wp}
\alias{boyd}
\title{Output formats for rmarkdown documents}
\usage{
wp(...)

boyd(...)
}
\arguments{
\item{\dots}{Arguments passed to \code{\link[bookdown]{pdf_document2}}.}
}
\value{
An R Markdown output format object.
}
\description{
Wrapper for \code{\link[bookdown]{pdf_document2}} to
produce documents in my preferred style.
}
\author{
Andrea Blasco
}
