% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{percent}
\alias{percent}
\title{Convert proportions in percentage strings with the \verb{\%} sign.}
\usage{
percent(x, format = "\%0.1f")
}
\arguments{
\item{x}{real number (proportion)}

\item{format}{to pass to the \code{sprintf} function}
}
\description{
Transforms proportions in strings with percent sign (with or without brackets).
}
\examples{
percent(0.15)
}
