% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{percent}
\alias{percent}
\title{Convert proportions in percentage strings with the \verb{\%} sign.}
\usage{
percent(x, digits = 0)
}
\arguments{
\item{x}{real number (proportion)}
}
\description{
Transforms proportions in strings with percent sign (with or without brackets).
}
\examples{
p <- 0.15
percent(0.15)
}
